src/main/resources/templates/springbootng/v2_v7/archetype/${projectName}-backend/scripts/keycloak/requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.4.4
PyYAML==5.2
requests==2.22.0
six==1.13.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
